Make some changes to memory alignments

+++ src/arm_7tdmi.h
@@ -119,7 +119,7 @@ class arm_7tdmi {
 
         // safely interface with memory
         u8 read_u8(u32);
-        u16 read_u16(u32);
+        u32 read_u16(u32);
         u32 read_u32(u32);
         void write_u8(u32, u8);
         void write_u16(u32, u16);
